[김평화] - 프로젝트 리스트, 팀원 평가 로직, 과업 달성률 로직, 헤더 프로필 이미지 사이즈 css 수정

+++ s20240130103/src/main/java/com/blackberry/s20240130103/kph/model/KphTask.java
@@ -1,14 +1,12 @@
 package com.blackberry.s20240130103.kph.model;
 
-import java.util.Date;
-
 import lombok.Data;
 
 @Data
 public class KphTask {
 	
-	private int project_no;
-	private int task_no;
+	private Long project_no;
+	private Long task_no;
 	private String task_title;
 	private String task_start;
 	private String task_end;
